Giving up on old method, will try masking next
Technique of re-scaling pixel values so they take up most of the
usable range (0, 255) is just making most of the pictures darker
or reducing contrast.  This may be because the underlying issue
of high peaks at the extremes is not actually addressed

Next, I will attempt to instead mask the brightest areas and
darken those as well as mask the darkest and brighten those.  It
is unclear how to do this while keeping the colors looking good
and the contrast hight but that is for future me to figure out